--- lib/multi-account-amp-monitoring/IAM Roles.json
@@ -1,129 +1,0 @@
-// AMP Workloads Account
-// Trust Relationship
-{
-    "Version": "2012-10-17",
-    "Statement": [
-        {
-            "Effect": "Allow",
-            "Principal": {
-                "AWS": "arn:aws:iam::${AMG_ACCOUNT_ID}:root"
-            },
-            "Action": "sts:AssumeRole",
-            "Condition": {}
-        }
-    ]
-}
-// Policy
-{
-    "Version": "2012-10-17",
-    "Statement": [
-        {
-            "Effect": "Allow",
-            "Action": [
-                "aps:ListWorkspaces",
-                "aps:DescribeWorkspace",
-                "aps:QueryMetrics",
-                "aps:GetLabels",
-                "aps:GetSeries",
-                "aps:GetMetricMetadata"
-            ],
-            "Resource": "*"
-        }
-    ]
-}
-
-
-// cloud-watch Workload account
-// Trust Relationship
-{
-    "Version": "2012-10-17",
-    "Statement": [
-        {
-            "Effect": "Allow",
-            "Principal": {
-                "AWS": "arn:aws:iam::${AMG_ACCOUNT_ID}:root"
-            },
-            "Action": "sts:AssumeRole",
-            "Condition": {}
-        }
-    ]
-}
-
-
-//Policy
-{
-  "Version": "2012-10-17",
-  "Statement": [
-    {
-      "Sid": "AllowReadingMetricsFromCloudWatch",
-      "Effect": "Allow",
-      "Action": [
-        "cloudwatch:DescribeAlarmsForMetric",
-        "cloudwatch:DescribeAlarmHistory",
-        "cloudwatch:DescribeAlarms",
-        "cloudwatch:ListMetrics",
-        "cloudwatch:GetMetricStatistics",
-        "cloudwatch:GetMetricData"
-      ],
-      "Resource": "*"
-    },
-    {
-      "Sid": "AllowReadingLogsFromCloudWatch",
-      "Effect": "Allow",
-      "Action": [
-        "logs:DescribeLogGroups",
-        "logs:GetLogGroupFields",
-        "logs:StartQuery",
-        "logs:StopQuery",
-        "logs:GetQueryResults",
-        "logs:GetLogEvents"
-      ],
-      "Resource": "*"
-    },
-    {
-      "Sid": "AllowReadingTagsInstancesRegionsFromEC2",
-      "Effect": "Allow",
-      "Action": ["ec2:DescribeTags", "ec2:DescribeInstances", "ec2:DescribeRegions"],
-      "Resource": "*"
-    },
-    {
-      "Sid": "AllowReadingResourcesForTags",
-      "Effect": "Allow",
-      "Action": "tag:GetResources",
-      "Resource": "*"
-    }
-  ]
-}
-
-
-
-amg-account
-//Trust Relationship
-{
-  "Version": "2012-10-17",
-  "Statement": [
-  {
-      "Effect": "Allow",
-      "Principal": {
-        "Service": "grafana.amazonaws.com"
-      },
-      "Action": "sts:AssumeRole"
-    }
- ]
-}
-//Policy
-{
-    "Version": "2012-10-17",
-    "Statement": [
-    {
-        "Effect": "Allow",
-        "Action": "sts:AssumeRole",
-        "Resource": "arn:aws:iam::${WLDB_ACCOUNT_ID}:role/AMGCloudWatchDataSourceRole"
-    },
-    {
-        "Effect": "Allow",
-        "Action": "sts:AssumeRole",
-        "Resource": "arn:aws:iam::${WLDA_ACCOUNT_ID}:role/AMGPrometheusDataSourceRole"
-    }
-   ]
-}
