+++ lib/multi-account-amp-monitoring/cloudwatch-iam-setup.ts
@@ -1,0 +1,50 @@
+import { NestedStack, NestedStackProps } from 'aws-cdk-lib';
+import * as blueprints from '@aws-quickstart/eks-blueprints';
+import { Construct } from 'constructs';
+import * as iam from 'aws-cdk-lib/aws-iam';
+import * as cdk from 'aws-cdk-lib';
+
+export class CloudWatchIamSetupStack extends NestedStack {
+
+    public static builder(roleName: string, trustAccount: string): blueprints.NestedStackBuilder {
+        return {
+            build(scope: Construct, id: string, props: NestedStackProps) {
+                return new CloudWatchIamSetupStack(scope, id, props, roleName, trustAccount);
+            }
+        };
+    }
+
+    constructor(scope: Construct, id: string, props: NestedStackProps, roleName: string, trustAccount: string) {
+        super(scope, id, props);
+
+        const role = new iam.Role(this, 'cloudwatch-iam-trust-role', {
+            roleName: roleName,
+            assumedBy: new iam.AccountPrincipal(trustAccount),
+            description: 'CloudWatch role to assume from central account',
+        });
+
+        role.addToPolicy(new iam.PolicyStatement({
+            actions: [
+                "cloudwatch:DescribeAlarmsForMetric",
+                "cloudwatch:DescribeAlarmHistory",
+                "cloudwatch:DescribeAlarms",
+                "cloudwatch:ListMetrics",
+                "cloudwatch:GetMetricStatistics",
+                "cloudwatch:GetMetricData",
+                "logs:DescribeLogGroups",
+                "logs:GetLogGroupFields",
+                "logs:StartQuery",
+                "logs:StopQuery",
+                "logs:GetQueryResults",
+                "logs:GetLogEvents",
+                "ec2:DescribeTags",
+                "ec2:DescribeInstances",
+                "ec2:DescribeRegions",
+                "tag:GetResources"
+            ],
+            resources: ["*"],
+        }));
+
+        new cdk.CfnOutput(this, 'CloudWatchTrustRole', { value: role ? role.roleArn : "none" });
+    }
+}
